scripts/adventofcode/pyproject.toml
[tool.poetry]
name = "aoc"
version = "0.1.0"
description = ""
authors = ["sugyan <sugi1982@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.1"
pillow = "^9.4.0"
requests = "^2.28.1"
cssutils = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
mypy = "^0.991"
types-requests = "^2.28.11.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
